starting redux development
