+++ src/store/constants/action-types.js
@@ -1,0 +1,1 @@
+export const CHANGE_STAT = "CHANGE_STAT"
